+++ packages/evokit-body/__tests__/Body.indent.test.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+import { Body } from '../src';
+
+describe('<Body body-indent="..." />', () => {
+    it('props body-indent', () => {
+        const wrapper = shallow(<Body body-indent='value' />);
+        expect(wrapper).toMatchSnapshot();
+
+        expect(wrapper.html()).toEqual('<body class="ek-body ek-body_indent_value"></body>');
+    });
+});