+++ packages/evokit-body/src/index.js
@@ -1,12 +1,13 @@
 import { createBlock } from 'evokit';
 
 export const Body = createBlock('body', 'body', [
-    'size',
-    'indent',
     'background',
     'color',
+    'display',
+    'indent',
+    'size',
 ]);
 
-export const BodySection = createBlock('div', 'body__section');
+export const BodySection = createBlock('div', 'body__section', ['display']);
 
 Body.Section = BodySection;